types: extract JSON primitive types to json.ts
